// BillingAndSubscriptionUI/src/app/models/payment.ts
export interface IPaymentPayload {
  subscriptionId: number;
  amount: number;
  paymentDate: string;
  paymentStatus: number;
}

export interface IPaymentMessage {
  message: string;
}

export interface IOverduePayments {
  id: string;
  subscriptionId: number;
  amount: number;
  paymentDate: string;
  paymentStatus: number;
}

export interface IOverduePaymentsResponse {
  message: string;
  overduePayments: IOverduePayments[];
}